Refactor policy DB into a PolicyDb struct in compact/policy_db.rs
Move SQLite read/write logic out of the binary and lib.rs into a PolicyDb
struct with open/write/lookup_action_values methods. Also adds step_interval
filtering (--step-interval CLI arg) to only save states at every Nth step.
